Prevent static libraries from being suggested (#7)
The "LIBRARY" config variable can be a static library. This is currently
handled by replacing the extension with the shared object extension for
the current OS. This makes an assumption about the possible name, so
instead we just filter out the name if it has a static object extension.

Additionally, the "LDLIBRARY" config variable may return an import
library on msys2 and mingw.