Add support for MSYS and MINGW32/64
On MINGW32/64, unlike on normal Windows installations, all libraries are
prefixed with 'lib'. MSYS does not prefix with lib.
MSYS is not detected as Windows, but follows it's convention of putting
DLLs in the same directory as the executable, so we add the path to the
executable to the search path.
MINGW32/64 and MSYS both define the DLLLIBRARY as the location of the
DLL for Python, while LDLIBRARY points to the import library. So we add
DLLLIBRARY to the list of possible names to check.