Change _linked_libpython_unix to check library handles given to it
The previous implementation only checked `ctypes.pythonapi`, which
depending on how the Python was configured could return a handle to the
executable or libpython. Attempts to filter out the executable and fall
back on other methods didn't work. Instead, we use the candidate names
to try loading the library and use dladdr to obtain the filepath.